\name{cuckoosearch}
\alias{cuckoosearch}
\title{Cuckoo Search metaheuristic algorithm to perform global.}
\usage{cuckoosearch(n = 25, maxIter = 10^5, pa = 0.25, Tol = 1e-05, nd = 2, lb = 0, ub = pi)}
\description{
	The Cuckoo Search function contains algorithm to get the minimum global of other optimization functions.
}
\examples{
	cuckoosearch()
}

\references{
	X.-S. Yang, S. Deb, \emph{Engineering Optimisation by Cuckoo Search}. Int. J. Mathematical Modelling and Numerical Optimisation, Vol. 1, No. 4, pp. 330–343, 2010.

	X.-S. Yang, S. Deb, \emph{Cuckoo search via L´evy flights,}. India. IEEE Publications, USA, pp. 210-214, 2009.
}